requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.1.0=pypi_0
addict=2.4.0=pypi_0
aliyun-python-sdk-core=2.15.0=pypi_0
aliyun-python-sdk-kms=2.16.2=pypi_0
anyio=3.7.1=pypi_0
argon2-cffi=23.1.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
attrs=23.2.0=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.12.3=pypi_0
black=23.3.0=pypi_0
blas=1.0=mkl
bleach=6.0.0=pypi_0
brotlipy=0.7.0=py37h27cfd23_1003
bzip2=1.0.8=h5eee18b_5
ca-certificates=2024.3.11=h06a4308_0
cachetools=5.3.3=pypi_0
certifi=2022.12.7=py37h06a4308_0
cffi=1.15.1=pypi_0
chardet=5.2.0=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
colorama=0.4.6=pypi_0
comm=0.1.4=pypi_0
crcmod=1.7=pypi_0
cryptography=42.0.5=pypi_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.11.0=pypi_0
debugpy=1.7.0=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
descartes=1.1.0=pypi_0
entrypoints=0.4=pypi_0
exceptiongroup=1.2.0=pypi_0
fastjsonschema=2.19.1=pypi_0
ffmpeg=4.3=hf484d3e_0
fire=0.6.0=pypi_0
flake8=3.9.2=pypi_0
fonttools=4.38.0=pypi_0
freetype=2.12.1=h4a9f257_0
giflib=5.2.1=h5eee18b_3
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
google-auth=2.29.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.62.1=pypi_0
idna=3.6=pypi_0
imageio=2.31.2=pypi_0
importlib-metadata=6.7.0=pypi_0
importlib-resources=5.12.0=pypi_0
iniconfig=2.0.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.16.2=pypi_0
ipython=7.34.0=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=8.1.2=pypi_0
jedi=0.19.1=pypi_0
jinja2=3.1.3=pypi_0
jmespath=0.10.0=pypi_0
joblib=1.3.2=pypi_0
jpeg=9e=h5eee18b_1
jsonschema=4.17.3=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=7.4.9=pypi_0
jupyter-console=6.6.3=pypi_0
jupyter-core=4.12.0=pypi_0
jupyter-server=1.24.0=pypi_0
jupyterlab-pygments=0.2.2=pypi_0
jupyterlab-widgets=3.0.10=pypi_0
kiwisolver=1.4.5=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
lerc=3.0=h295c915_0
libdeflate=1.17=h5eee18b_1
libedit=3.1.20230828=h5eee18b_0
libffi=3.2.1=hf484d3e_1007
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
llvmlite=0.31.0=pypi_0
lyft-dataset-sdk=0.0.8=pypi_0
lz4-c=1.9.4=h6a678d5_0
markdown=3.4.4=pypi_0
markdown-it-py=2.2.0=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.5.2=pypi_0
matplotlib-inline=0.1.6=pypi_0
mccabe=0.6.1=pypi_0
mdurl=0.1.2=pypi_0
mistune=3.0.2=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py37h7f8727e_0
mkl_fft=1.3.1=py37hd3c417c_0
mkl_random=1.2.2=py37h51133e4_0
mmcv-full=1.4.0=pypi_0
mmdet=2.19.0=pypi_0
mmdet3d=0.18.1=dev_0
mmsegmentation=0.20.0=pypi_0
model-index=0.1.11=pypi_0
mypy-extensions=1.0.0=pypi_0
nbclassic=1.0.0=pypi_0
nbclient=0.7.4=pypi_0
nbconvert=7.6.0=pypi_0
nbformat=5.8.0=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pypi_0
nettle=3.7.3=hbbd107a_1
networkx=2.2=pypi_0
notebook=6.5.6=pypi_0
notebook-shim=0.2.4=pypi_0
numba=0.48.0=pypi_0
numpy=1.19.5=pypi_0
nuscenes-devkit=1.1.10=pypi_0
oauthlib=3.2.2=pypi_0
opencv-python=4.9.0.80=pypi_0
opendatalab=0.0.10=pypi_0
openh264=2.1.1=h4ff587b_0
openmim=0.3.9=pypi_0
openssl=1.0.2u=h7b6447c_0
openxlab=0.0.10=pypi_0
ordered-set=4.1.0=pypi_0
oss2=2.17.0=pypi_0
packaging=24.0=pypi_0
pandas=1.1.5=pypi_0
pandocfilters=1.5.1=pypi_0
parso=0.8.3=pypi_0
pathspec=0.11.2=pypi_0
pexpect=4.9.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.4.0=py37h6a678d5_0
pip=22.3.1=py37h06a4308_0
pkgutil-resolve-name=1.3.10=pypi_0
platformdirs=4.0.0=pypi_0
plotly=5.18.0=pypi_0
pluggy=1.2.0=pypi_0
plyfile=0.9=pypi_0
prettytable=3.7.0=pypi_0
prometheus-client=0.17.1=pypi_0
prompt-toolkit=3.0.43=pypi_0
protobuf=3.20.3=pypi_0
psutil=5.9.8=pypi_0
ptyprocess=0.7.0=pypi_0
pyasn1=0.5.1=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycocotools=2.0.7=pypi_0
pycodestyle=2.7.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pycryptodome=3.20.0=pypi_0
pyflakes=2.3.1=pypi_0
pygments=2.17.2=pypi_0
pyopenssl=23.0.0=py37h06a4308_0
pyparsing=3.1.2=pypi_0
pyquaternion=0.9.9=pypi_0
pyrsistent=0.19.3=pypi_0
pysocks=1.7.1=py37_1
pytest=7.4.4=pypi_0
python=3.7.0=h6e4f718_3
python-dateutil=2.9.0.post0=pypi_0
pytorch-mutex=1.0=cuda
pytz=2023.4=pypi_0
pywavelets=1.3.0=pypi_0
pyyaml=6.0.1=pypi_0
pyzmq=24.0.1=pypi_0
qtconsole=5.4.4=pypi_0
qtpy=2.4.1=pypi_0
readline=7.0=h7b6447c_5
requests=2.28.2=pypi_0
requests-oauthlib=1.4.0=pypi_0
rich=13.7.1=pypi_0
rsa=4.9=pypi_0
scikit-image=0.19.3=pypi_0
scikit-learn=1.0.2=pypi_0
scipy=1.7.3=pypi_0
send2trash=1.8.2=pypi_0
setuptools=60.2.0=pypi_0
shapely=1.8.5=pypi_0
six=1.16.0=pyhd3eb1b0_1
sniffio=1.3.1=pypi_0
soupsieve=2.4.1=pypi_0
spconv=1.2.1=pypi_0
sqlite=3.33.0=h62c20be_0
tabulate=0.9.0=pypi_0
tenacity=8.2.3=pypi_0
tensorboard=2.11.2=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
termcolor=2.3.0=pypi_0
terminado=0.17.1=pypi_0
terminaltables=3.1.10=pypi_0
threadpoolctl=3.1.0=pypi_0
tifffile=2021.11.2=pypi_0
tinycss2=1.2.1=pypi_0
tk=8.6.12=h1ccaba5_0
tomli=2.0.1=pypi_0
torch=1.9.1+cu111=pypi_0
torch-scatter=2.0.8=pypi_0
torchaudio=0.9.1=pypi_0
torchvision=0.10.1+cu111=pypi_0
tornado=6.2=pypi_0
tqdm=4.65.2=pypi_0
traitlets=5.9.0=pypi_0
trimesh=2.35.39=pypi_0
typed-ast=1.5.5=pypi_0
typing-extensions=4.7.1=pypi_0
typing_extensions=4.3.0=py37h06a4308_0
urllib3=1.26.18=pypi_0
wcwidth=0.2.13=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.6.1=pypi_0
werkzeug=2.2.3=pypi_0
wheel=0.38.4=py37h06a4308_0
widgetsnbextension=4.0.10=pypi_0
xz=5.4.6=h5eee18b_0
yapf=0.40.1=pypi_0
zipp=3.15.0=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
